% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_assess.R
\name{cn_addAUPRs}
\alias{cn_addAUPRs}
\title{utility function to compile cn_calcPRs}
\usage{
cn_addAUPRs(bigList)
}
\arguments{
\item{bigList}{alist}
}
\value{
something
}
\description{
utility function to compile cn_calcPRs
}
