% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNet.R
\name{cn_apply}
\alias{cn_apply}
\title{Apply CellNet to query data}
\usage{
cn_apply(expQuery, stQuery, cnProc, dLevelQuery = "description1")
}
\arguments{
\item{expQuery}{matrix of expression values}

\item{stQuery}{data.frame expression meta data}

\item{cnProc}{CellNet object result of running cn_make_processor}

\item{develLevelQuery}{column name in stQuery to grou query samples for grn status plotting}
}
\value{
cnRes object
}
\description{
Classifies query data and computes GRN status
}
\examples{
cnRes<-cn_apply(expQuery, stQuery, cnProc, "description1")

}
