% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_assess.R
\name{cn_assessGRN}
\alias{cn_assessGRN}
\title{assess performance of GRN predictions based on zscores. this function runs cn_calcPRs on all TFs in gol_standard}
\usage{
cn_assessGRN(zscores, goldStandard, tRange = seq(3, 6, by = 0.25),
  random = FALSE, randomIter = 10, funType = "intersect")
}
\arguments{
\item{zscores}{matrix of zscores}

\item{goldStandard}{gold stanard gene list}

\item{tRange}{threshold sequence}

\item{random}{boolean, whether to compute a random selection of calls, too, based on the number of genes actually predicted}

\item{randomIter}{numer of random selects to make, returns only the average}

\item{funType}{how to treat >1 TF, intersect or union}

\item{universe}{genes tested}
}
\value{
List of performance results of form: data frame of Score (cutoff), Precision, Recall, Predictions (n), obs = real GRN, rand = list of random GRNs, one per
}
\description{
Recall (Sensitivity): fraction of gold standard regulatory relationships detected by GRN. Precision (1-FPR): Proportion of called relationships that are in the gold standard
}
\examples{
prList<-cn_assessGRN(zscores, gs, random=TRUE)

}
