% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_assess.R
\name{cn_calcPRs}
\alias{cn_calcPRs}
\title{assess performance of GRN predictions based on zscores}
\usage{
cn_calcPRs(zscores, tfs, goldStandard, universe = NULL, tRange = seq(3, 6,
  by = 0.25), random = FALSE, randomIter = 10, funType = "intersect")
}
\arguments{
\item{zscores}{matrix of zscores}

\item{tfs}{vector of tf(s) perturbed}

\item{goldStandard}{gold stanrd gene list}

\item{universe}{genes tested}

\item{tRange}{threshold sequence}

\item{random}{boolean, whether to compute a random selection of calls, too, based on the number of genes actually predicted}

\item{randomIter}{numer of random selects to make, returns only the average}

\item{funType}{how to treat >1 TF, intersect or union}
}
\value{
List of performance results of form: data frame of Score (cutoff), Precision, Recall, Predictions (n), obs = real GRN, rand = list of random GRNs
}
\description{
Recall (Sensitivity): fraction of gold standard regulatory relationships detected by GRN. Precision (1-FPR): Proportion of called relationships that are in the gold standard
}
\examples{
grnPR<-cn_calcPRs(zscs, "Sox2", aGoldStandard[["Sox2"]], allGenes, tRange=seq(0,8, by=.5), random=TRUE, randomIter=3);

}
