% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{cn_class_perf_hm_ordered}
\alias{cn_class_perf_hm_ordered}
\title{Order classifier performance heatmap by description1 label of validation samples}
\usage{
cn_class_perf_hm_ordered(cnProc, classifierPerformance)
}
\arguments{
\item{cnProc}{CellNet Processor object from training}

\item{classifierPerformance}{Classifier performance object generated with cn_splitMakeAssess}
}
\description{
Order classifier performance heatmap by description1 label of validation samples
}
