% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_clr}
\alias{cn_clr}
\title{make a CLR-like matrix for assessment purposes}
\usage{
cn_clr(sampTab, expDat, species = "Mm", tfs = NA, grnSampSize = 40,
  normDat = FALSE, dLevel = "description1")
}
\arguments{
\item{expDat}{properly normalized expression matrix}

\item{tfs}{vector of transcription factors}

\item{grnSampSize}{min number of samples per ct or group}

\item{normDat}{whether to quantile normalize the data prior to computing correlations}

\item{dLevel}{column name in sample table to group samples by}

\item{samptTab}{sample table}
}
\value{
zscore matrix
}
\description{
computes gene-gene correlation, then clr-like zscores
}
