% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_assess.R
\name{cn_compileGRNassessments}
\alias{cn_compileGRNassessments}
\title{compile results of assessing GRNs for multiple TFs}
\usage{
cn_compileGRNassessments(prs)
}
\arguments{
\item{prs}{result of running cn_assessGRN}
}
\value{
data frame
}
\description{
result can be immediately plotted
}
\examples{
prList<-cn_assessGRN(zscores, gs, random=TRUE)
assessedGRN<-cn_compileGRNassessments(prList)
ggplot(assessedGRN, aes(x=factor(Score), y=Fold)) +
 geom_boxplot(alpha=.75,colour="black", outlier.colour="#BDBDBD", outlier.size=.5, fill="#FF9999") +
 theme_bw() + ggtitle("AUPR Fold increase") + ylim(0,15) + ylab("AUPR fold improvement over random") + 
 xlab("GRN Z-score threshold")

}
