% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_assess.R
\name{cn_convertPR}
\alias{cn_convertPR}
\title{convert a result of cn_calcPRs to something easily plot-able}
\usage{
cn_convertPR(prRes)
}
\arguments{
\item{prRes}{result of running cn_calcPRs}
}
\value{
something
}
\description{
convert a result of cn_calcPRs to something easily plot-able
}
