% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_extractRegsDF}
\alias{cn_extractRegsDF}
\title{extracts the TRs, zscores, and corr values passing thresh}
\usage{
cn_extractRegsDF(zscores, corrMatrix, genes, threshold)
}
\arguments{
\item{zscores, }{# zscore matrix, non-TFs already removed from columns}

\item{corrMatrix, }{# correlation matrix}

\item{genes, }{# vector of target genes}

\item{threshold}{# zscore threshold}
}
\value{
data.frame(target=targets, reg=regulators, zscore=zscoresX, corr=correlations);
}
\description{
extracts the TRs, zscores, and corr values passing thresh
}
