% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_get_targets_of}
\alias{cn_get_targets_of}
\title{get targets of tfs}
\usage{
cn_get_targets_of(aGraph, tfs)
}
\arguments{
\item{aGraph}{iGraph network}

\item{tfs}{vector of TF names}
}
\value{
list of tf-> targets
}
\description{
get targets of tfs
}
