% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_getRawGRN}
\alias{cn_getRawGRN}
\title{get raw GRN from zscores, and corr}
\usage{
cn_getRawGRN(zscores, corrs, targetGenes, zThresh = 4)
}
\arguments{
\item{zscores}{zscores matrix}

\item{corrs}{correlation matrix}

\item{targetGenes}{target genes}

\item{zThresh}{zscore threshold}
}
\value{
list of grnTable and corresponding graph
}
\description{
get raw GRN from zscores, and corr
}
