% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{cn_HmClass}
\alias{cn_HmClass}
\title{heatmap of the classification result}
\usage{
cn_HmClass(cnRes, isBig = FALSE)
}
\arguments{
\item{cnRes}{returned from cn_sapply}

\item{isBig}{is this a big heatmap? TRUE or FALSE}
}
\value{
nothing
}
\description{
Heatmap of the classification result.
}
\examples{
cn_HmClass(cnRes, isBig=TRUE)

}
