% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{cn_HmClass_disease}
\alias{cn_HmClass_disease}
\title{heatmap of gene expression for disease modeling studies
Added 6-4-18}
\usage{
cn_HmClass_disease(cnResQuery, col_annotations, isBig = TRUE)
}
\arguments{
\item{cnResQuery}{cellNet query results object}

\item{isBig}{boolean}
}
\value{
nothing
}
\description{
heatmap of gene expression for disease modeling studies
Added 6-4-18
}
\examples{
cnRes_iPSC <- utils_loadObject("cn_iPSC_pooled_results_nat_prot_Jul_14_2017.R")
col_annotations_iPSC <- data.frame(Disease_Status = factor(sampTab_iPSC$disease_status), Study = factor(sampTab_iPSC$study_id))
set.seed(0) # reseed random number generator, makes column coloring reproducible
cn_HmClass_disease(cnRes_iPSC, col_annotations_iPSC)

}
