% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{cn_HmVars}
\alias{cn_HmVars}
\title{heatmap of gene expression}
\usage{
cn_HmVars(expDat, sampTab, isBig = FALSE, dist = "correlation")
}
\arguments{
\item{expDat}{expression matrix}

\item{sampTab}{sample table}

\item{isBig}{boolean}

\item{dist}{distance metricx used, determines also whether to row-scale or not}
}
\value{
nothing
}
\description{
heatmap of gene expression
}
\examples{
selectedGenes<-c("Actb", "Gapdh", "Hlf", "Sox2")
cn_HmVars(expDat[selectedGenes,], sampTab)

}
