% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_make_grn}
\alias{cn_make_grn}
\title{Reconstruct CT-specific GRNs}
\usage{
cn_make_grn(sampTab, expDat, species = "Mm", tfs = NA, grnSampSize = 0,
  normDat = FALSE, corrs = NA, zscores = NA, cval = 0.5,
  cvalGK = 0.75, dLevel = "description1", dLevelGK = "description6",
  zThresh = 4, holmSpec = 1e-06, prune = FALSE)
}
\arguments{
\item{expDat}{properly normalized expression matrix}

\item{tfs}{vector of transcription factors}

\item{grnSampSize}{min number of samples per ct or group}

\item{normDat}{whether to quantile normalize the data prior to computing correlations}

\item{corrs}{matrix of gene-gene pearson correlations}

\item{zscores}{CLR-defined context dependent zscores}

\item{cval}{template matching threshold for overall CT specific expression}

\item{cvalGK}{template matching threshold for developmentally shared CT specific expression}

\item{dLevel}{column name in sample table to group samples by}

\item{holmSpec}{pvalue threshold for template matching}

\item{samptTab}{sample table}

\item{zthresh}{GRN zscore threshold at which to remove weak edges}
}
\value{
list of overallGRN (desc), specGenes (desc), ctGRNs (desc)
}
\description{
runs getRawGRN, findSpecGenes, and specGRNs
}
\examples{
system.time(grnAll<-cn_make_grn(stAll, expAll, species='Mm', zThresh=6) )

}
