% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNet.R
\name{cn_make_processor}
\alias{cn_make_processor}
\title{make a CellNet object}
\usage{
cn_make_processor(expTrain, stTrain, ctGRNs, dLevel = "description1",
  classWeight = TRUE, exprWeight = TRUE, sidCol = "sample_id")
}
\arguments{
\item{expTrain}{expression matrix}

\item{stTrain}{sample table}

\item{ctGRNs}{result of running cn_grnDoRock}

\item{dLevel}{sample table column name to operate on}

\item{classWeight}{weight GRN est by classification importance}

\item{exprWeight}{weight GRN est by gene expression}

\item{sidCol}{column in stTrain that contains unique id for each sample and is the colname in expTrain}
}
\value{
cnProc
}
\description{
Does lots of stuff, but mostly you will care about the fact that it trains C/T classifiers
}
\examples{


}
