% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_make_tVals_predict}
\alias{cn_make_tVals_predict}
\title{cn_make_tVals_predict}
\usage{
cn_make_tVals_predict(expDat, sampTab, dLevel = "description1")
}
\arguments{
\item{expDat}{training data}

\item{sampTab}{training sample table}

\item{dLevel="description1"}{column to define CT}
}
\value{
tVals list of ct->mean->named vector of average gene expression, ->sd->named vector of gene standard deviation
}
\description{
predicts SD based on mean expression
}
