% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNet.R
\name{cn_makeClassifier}
\alias{cn_makeClassifier}
\title{make a single RF classifier}
\usage{
cn_makeClassifier(trainScores, ctt, stTrain, dLevel = "description1")
}
\arguments{
\item{trainScores}{numeric matrix of predictors}

\item{ctt}{what cell type/tissues is this?}

\item{stTrain}{sample table}
}
\value{
randomForest classifier
}
\description{
uses R's randomForest package
}
