% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNet.R
\name{cn_makeRFs}
\alias{cn_makeRFs}
\title{make classifiers}
\usage{
cn_makeRFs(expTrain, stTrain, geneLists, dLevel = "description1")
}
\arguments{
\item{expTrain}{expression matrix}

\item{stTrain}{sample table}

\item{dLevel}{stTrain column name to split on}

\item{list}{of dLevel=>vector of genes}
}
\value{
list of classifiers
}
\description{
Make one Random Forest classification per unique dLevel
}
