% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_makeSGs}
\alias{cn_makeSGs}
\title{make induced subgraphs from gene lists and iGraph object}
\usage{
cn_makeSGs(bigIG, geneLists)
}
\arguments{
\item{bigIG}{super-graph}

\item{geneLists}{gene lists}
}
\value{
list of graphs
}
\description{
make induced subgraphs from gene lists and iGraph object
}
