% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_MakeTLs}
\alias{cn_MakeTLs}
\title{get targets of tFs}
\usage{
cn_MakeTLs(graphList)
}
\arguments{
\item{graphList}{a list of networks represented as iGraphs}
}
\value{
list of tf=>targets
}
\description{
get targets of tFs
}
