% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_nis}
\alias{cn_nis}
\title{network influence score}
\usage{
cn_nis(cnRes, cnProc, subnet, ctt, relaWeight = 1)
}
\arguments{
\item{cnRes}{object result of running cn_apply}

\item{cnProc}{object result of running cn_make_processor}

\item{subnet}{name of subnet to evaluage}

\item{ctt}{string indicating the CT to compare against}
}
\value{
numeric matrix where rows are TFs in CT GRN and columns are query samples
}
\description{
Computes network influence score (NIS). See paper for details.
}
