% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_nis_all}
\alias{cn_nis_all}
\title{Network Influence Score for all GRNs}
\usage{
cn_nis_all(cnRes, cnProc, ctt, relaWeight = 1)
}
\arguments{
\item{cnRes}{object result of running cn_apply}

\item{cnProc}{object result of running cn_make_processor}

\item{ctt}{string indicating the CT to compare against}

\item{relaWeight}{whether to weight by overall expression such that TFs with higher expression in ctt are more important (1=do the weighting)}
}
\value{
list of numeric matrix of TF scores
}
\description{
Runs cn_nis on all GRNs
}
