% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_normalizeScores}
\alias{cn_normalizeScores}
\title{Normalize grn status as compared to training data}
\usage{
cn_normalizeScores(ctrlScores, queryScores, subNets)
}
\arguments{
\item{ctrlScores}{a list of subnet->mean value, all subnets}

\item{queryScores}{a matrix, rownames = subnet names, all subnets}

\item{subNets}{a vector of subnets names to use}
}
\value{
normalized grn status matrix
}
\description{
Divide the query scores by the mean values in the training data.
}
