% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNet.R
\name{cn_outputRes}
\alias{cn_outputRes}
\title{Write out classification scores, normalized data to csv}
\usage{
cn_outputRes(cnRes, tfScores, prefix)
}
\arguments{
\item{cnRes}{result of running cn_apply}

\item{tfScores}{network inlfuence scores}

\item{prefix}{filename prefix}
}
\value{
filename
}
\description{
and it compresses them, too
}
