% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNet.R
\name{cn_remake_processor}
\alias{cn_remake_processor}
\title{make a CellNet object from an existing one with subset of genes and/or new expTrain}
\usage{
cn_remake_processor(cnProc, newGenes = NA, expTrain = NA,
  sidCol = "sample_id")
}
\arguments{
\item{cnProc}{old cnProc}

\item{newGenes}{optional gene subset to use}

\item{expTrain}{optional new expTrain}

\item{sidCol}{column in stTrain that contains unique id for each sample and is the colname in expTrain}

\item{ctGRNs}{result of running cn_grnDoRock}

\item{dLevel}{sample table column name to operate on}

\item{classWeight}{weight GRN est by classification importance}

\item{exprWeight}{weight GRN est by gene expression}
}
\value{
cnProc
}
\description{
Does lots of stuff, but mostly you will care about the fact that it trains C/T classifiers
}
\examples{


}
