% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{cn_s3_fetchFastq}
\alias{cn_s3_fetchFastq}
\title{Fetch fastq files from S3}
\usage{
cn_s3_fetchFastq(bucket, path, sampTab, fnameCol = "fname", compressed = NA)
}
\arguments{
\item{bucket}{name of bucket}

\item{path}{path to fastqs}

\item{sampTab}{sample table with a column fname listing the fastq files}

\item{fnameCol}{name of column listing fastq files}
}
\value{
sammple table with de-compressed file names replacing old file names
}
\description{
It assumes that the files are publicly accessible. If not then you should use s3_get after you have run aws configure and enter your aws credentials
}
