% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{cn_salmon}
\alias{cn_salmon}
\title{Derive gene expression estimates compatible with CellNet}
\usage{
cn_salmon(sampTab, total = 1e+05, fnameCol = "fname", finalLength = 40,
  delOrig = FALSE, salmonIndex = "MM_GRCh38.SalmonIndex.030816",
  geneTabfname = "geneToTrans_Mus_musculus.GRCm38.80.exo_Jun_02_2015.R",
  refDir = "/media/ephemeral1/dat/ref/",
  salmonPath = "~/rnaseq/SalmonBeta-0.6.1_DebianSqueeze/bin")
}
\arguments{
\item{sampTab}{sample table with fname column point to fastq files}

\item{total}{number of reads to normalzie to}

\item{fnameCol}{column name of fastq files}

\item{finalLength}{length of reads after trimming}

\item{species}{mouse or human}

\item{bucket}{where to get them}

\item{what}{path on s3 to get them}
}
\value{
nothing
}
\description{
Derive gene expression estimates compatible with CellNet
}
