% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_sampR_to_pattern}
\alias{cn_sampR_to_pattern}
\title{return a pattern for use in cn_testPattern (template matching)}
\usage{
cn_sampR_to_pattern(sampR)
}
\arguments{
\item{sampR}{vector}
}
\value{
ans
}
\description{
return a pattern for use in cn_testPattern (template matching)
}
