% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{cn_setup}
\alias{cn_setup}
\title{set up directories on ephemeral drives that will be needed to store large files}
\usage{
cn_setup(local = FALSE, wrDir = "/media/ephemeral0/analysis")
}
\arguments{
\item{wrDir}{working directory}
}
\value{
nothing
}
\description{
Creates a dir /media/ephemeral0/analysis, changes ownership ti ec2-user, sets cwd to this dir, sets up dir in /media/ephemeral1 for reference data
}
