% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_specGenesAll}
\alias{cn_specGenesAll}
\title{finds general and context dependent specifc genes}
\usage{
cn_specGenesAll(expDat, sampTab, holm = 1e-10, cval = 0.5, cvalGK = 0.75,
  dLevel = "description1", dLevelGK = "description2", prune = FALSE)
}
\arguments{
\item{expDat}{expression matrix}

\item{sampTab}{sample table}

\item{holm}{pvalue threshold for template matching}

\item{cval}{template matching threshold for overall CT specific expression}

\item{cvalGK}{template matching threshold for developmentally shared CT specific expression}

\item{dLevel}{"description1",}

\item{dLevelGK}{"description2"}
}
\value{
list of $matcher${cell_type}->{germ_layer}$context$general${cell_type}->gene vector etc
}
\description{
finds general and context dependent specifc genes
}
