% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_specGRNs}
\alias{cn_specGRNs}
\title{extract sub-networks made up of CT genes;}
\usage{
cn_specGRNs(rawGRNs, specGenes)
}
\arguments{
\item{specGenes}{result of running cn_specGenesAll}

\item{rawGRNsresult}{of running cn_getRawGRN}
}
\value{
list(geneLists=geneLists, graphLists=graphLists, tfTargets=tfTargets)
}
\description{
extract sub-networks made up of CT genes;
}
