% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNet.R
\name{cn_splitMakeAssess}
\alias{cn_splitMakeAssess}
\title{make classifiers and assess}
\usage{
cn_splitMakeAssess(stDat, expDat, grns, prop = 0.5, dLevel = "description1",
  dLevelStudy = "exp_id", dLevelSID = "sample_id")
}
\arguments{
\item{stDat}{sample table}

\item{expDat}{expression data}

\item{grns}{result of running cn_grnDoRock}

\item{prop}{numeric 0<prop<1 what proportion of data to train classifiers on}

\item{dLevel}{sample table column to use as response value in classifier (e.g. cell type)}

\item{dLevelStudy}{column name to indicate experiment or study id}

\item{dLevelSID}{column to indicate sample}
}
\value{
list if classifiers, classifier result as applied to held out data, ROCs
}
\description{
split the data by prop, make classifiers, assess them on held out data
}
\examples{
assessmentResults<-cn_splitMakeAssess(stTrain, expTrain, ctGRNs, prop=.5, dLevel="description1")

}
