% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_assess.R
\name{cn_summ_AUPRs}
\alias{cn_summ_AUPRs}
\title{processes result of cn_addAUPRs: p-value of fold improvements,df for visualization}
\usage{
cn_summ_AUPRs(bigList, name)
}
\arguments{
\item{bigList}{result of cn_addAUPRs}

\item{name}{name of analysis}
}
\value{
something
}
\description{
processes result of cn_addAUPRs: p-value of fold improvements,df for visualization
}
