% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{cn_tfScore}
\alias{cn_tfScore}
\title{transcription factor score}
\usage{
cn_tfScore(cnProc, expDat, ctt, tfnames = NULL)
}
\arguments{
\item{cnProc}{object result of running cn_make_processor}

\item{ctt}{string indicating the CT to compare against}

\item{cnRes}{object result of running cn_apply}

\item{tfname}{transcription factor}
}
\value{
numeric matrix where rows are TFs and columns are query samples, values are zscores
}
\description{
Computes TF score (TFS). See forthcoming paper for details.
}
