% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNet.R
\name{cn_writeReadme}
\alias{cn_writeReadme}
\title{Helper function to write a markdown line for the Readme}
\usage{
cn_writeReadme(cnProc, species, date = NA,
  url = "https://s3.amazonaws.com/CellNet/rna_seq/")
}
\arguments{
\item{cnProc}{cnProc}

\item{species}{species}

\item{date}{date made}

\item{url}{url}
}
\value{
string
}
\description{
This writes a string that includes the number of samples, the name and number of C/Ts, and S3 links for a given cnProc
}
\examples{
myMDstring<-cn_writeReadme(cnProc, "mouse", date="Oct_25_2016")

}
