% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{expr_readSampTab}
\alias{expr_readSampTab}
\title{properly read a csv file as a sampTab}
\usage{
expr_readSampTab(fname)
}
\arguments{
\item{fname}{csv file}
}
\value{
data frame, should have sample_id, sample_name, description1, description2, exp_id, file_name columns
}
\description{
properly read a csv file as a sampTab
}
