% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{expr_reorder}
\alias{expr_reorder}
\title{re-order the sampTab, by dLevel, given dLevel names}
\usage{
expr_reorder(sampTab, dLevel, nnames)
}
\arguments{
\item{sampTab}{sample table}

\item{dLevel}{col to re-order on}

\item{nnames}{lelves to reorder}
}
\value{
new sample table
}
\description{
re-order the sampTab, by dLevel, given dLevel names
}
