% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{fastq_trim}
\alias{fastq_trim}
\title{Trim reads to specified length}
\usage{
fastq_trim(sampTab, fnameCol = "fname", finalLength = 40, outDir = "./")
}
\arguments{
\item{sampTab}{sample table}

\item{fnameCol}{sampTab column with name of fastq files to trim}

\item{finalLength}{final length of each read post trimming}

\item{outDir}{where to store the trimmed fastq files}
}
\value{
sampTab with filenames of trimmed fastqs appended
}
\description{
Use cutadapt to trim reads from both ends to finalLength
}
