% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{feature_count_par}
\alias{feature_count_par}
\title{count the number of reads that overlap with genomic features}
\usage{
feature_count_par(sampTab, gtfpath, sname = "sra_id", target = "mouse")
}
\arguments{
\item{sampTab}{sample table}

\item{gtfpath}{gtf of genomic features}

\item{sname}{column to use to name output files}

\item{target}{species, e.g. 'mouse'}
}
\value{
sample table with hit rates appended
}
\description{
count the number of reads that overlap with genomic features uses samtools/htseq-count
}
