% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{fetch_salmon_indices}
\alias{fetch_salmon_indices}
\title{fetch files needed to run salmon}
\usage{
fetch_salmon_indices(destination = "/media/ephemeral1/dat/ref",
  species = "mouse", bucket = "cellnet-rnaseq", dir = "ref", iFile = NA)
}
\arguments{
\item{destination}{where to out the indices}

\item{species}{mouse or human}

\item{bucket}{where to get them}

\item{dir}{what path on s3 to get them}
}
\value{
nothing
}
\description{
fetch files needed to run salmon
}
