% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{fetchAndConvert}
\alias{fetchAndConvert}
\title{fetch sras and converts them to fastqs, can take a REALLY long time and is unreliable}
\usage{
fetchAndConvert(sampTab, tmpPath = "/media/ephemeral0/dat")
}
\arguments{
\item{sampTab}{sample table}

\item{tmpPath}{where sras are stored}
}
\value{
sample table with read length appended
}
\description{
fetch sras and converts them to fastqs, can take a REALLY long time and is unreliable. SRA toolkit must be installed and configured. Make sure the store the SRAs on a drive with lots of space. Must be consistent with tmpPath.
}
