% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{fetchIndexHandler}
\alias{fetchIndexHandler}
\title{wrapper to fetch files needed to run salmon in case of errors}
\usage{
fetchIndexHandler(destination = "/media/ephemeral1/dat/ref",
  species = "mouse", bucket = "cellnet-rnaseq", dir = "ref",
  iFile = "salmon.index.mouse.050316.tgz")
}
\arguments{
\item{destination}{where to put the indices}

\item{species}{mouse or human}

\item{bucket}{where to get them}

\item{dirw}{hat path on s3 to get them}

\item{salmonVersion}{version of salmon, if > 0.6 gets latest indices}
}
\value{
nothing
}
\description{
fetch files needed to run salmon
}
