% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{gene_expr_sum}
\alias{gene_expr_sum}
\title{Sum transcription expression estimates to gene-level expression measures}
\usage{
gene_expr_sum(expDatList, numCores = 10,
  geneTabfname = "/media/ephemeral1/dat/ref/geneToTrans_Mus_musculus.GRCm38.80.exo_Jun_02_2015.R",
  nameCol = "gene_name")
}
\arguments{
\item{expDatList}{result of running salmon_load_tranEst}

\item{numCores}{num of cores to use for parallel}

\item{geneTabfname}{gene <-> transcript mapping, df that needs cols: gene_id, transcript_id}

\item{nameCol}{gene ann table column name to average over}
}
\value{
list(TPM=ansTPM, counts=ansCounts);
}
\description{
Sums all trnascript level expression estimates to a single gene level estimate. Needs a table that maps transcript IDs to gene IDs.
}
