% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{get_grn_scores}
\alias{get_grn_scores}
\title{Internal function, subset GRN scores for given sample indices}
\usage{
get_grn_scores(sampTab, sample_indices, qScores)
}
\arguments{
\item{sampTab}{sample metadata table}

\item{sample_indices}{integer index vector of which samples to subset}

\item{qScores}{matrix of study GRN scores}
}
\description{
Internal function, subset GRN scores for given sample indices
}
