% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{grn_corr_round}
\alias{grn_corr_round}
\title{gene-gene correlations, and round}
\usage{
grn_corr_round(expDat)
}
\arguments{
\item{expDat}{expression matrix}
}
\value{
correlation matrix
}
\description{
gene-gene correlations, and round
}
