% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{grn_zscores}
\alias{grn_zscores}
\title{compute CLR-like zscores}
\usage{
grn_zscores(corrVals, tfs)
}
\arguments{
\item{corrVals}{correlation matrix}

\item{tfs}{vector of  transcriptional regualtor names}
}
\value{
zscore matrix
}
\description{
compute CLR-like zscores
}
