% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{hisat_findIndices}
\alias{hisat_findIndices}
\title{finds hisat indecies}
\usage{
hisat_findIndices(dir = "/media/ephemeral1/dat/ref/hisat2Indices")
}
\arguments{
\item{dir}{base path to indecies}
}
\value{
df of nickname -> path
}
\description{
finds hisat indecies
}
