% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{hisat_QC_align}
\alias{hisat_QC_align}
\title{performs qc on fastq file}
\usage{
hisat_QC_align(fastq, indexDF, sname, target = "mouse",
  hisatDir = "~/rnaseq/hisat2-2.0.3-beta")
}
\arguments{
\item{fastq}{file name}

\item{indexDF}{index data frame result of running hisat_findIndices}

\item{sname}{col name}

\item{target}{species/genome}

\item{hisatDir}{path to hisat}
}
\value{
sampTab with hisat QC stats added AND features counts vs target genome
}
\description{
aligns fastq to genomes indexed and pointed to by indexDF, counts features aligning to target
}
