% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsTFs}
\alias{hsTFs}
\title{human transcriptional regulators}
\format{A vector with 1999 gene names
\describe{ 
}}
\source{
org.Hs.eg.db and GO.db
}
\usage{
hsTFs
}
\description{
A vector of gene symbols representing transcription factors, defined as genes annotated with GO BP as 'regulation of transcription' and 'DNA binding'
}
\keyword{datasets}

