% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{htseqCount}
\alias{htseqCount}
\title{compiles read count hits}
\usage{
htseqCount(df, promNames = c("rRNA", "Mt_rRNA", "protein_coding",
  "__no_feature", " __ambiguous", "__too_low_aQual", "__not_aligned",
  "__alignment_not_unique"))
}
\arguments{
\item{df}{data.frame that results from htseq-count}

\item{promNames}{vector of features names of interest}
}
\value{
data.frame of feature, proportion of reads mapping to this feature
}
\description{
compiles read count hits
}
