% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{mat_zscores}
\alias{mat_zscores}
\title{compute context dependent zscores}
\usage{
mat_zscores(corrMat)
}
\arguments{
\item{corrMat}{correlation matrix}
}
\value{
matrix of clr zscores
}
\description{
slightly modidied from JJ Faith et al 2007
}
