% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{minDif}
\alias{minDif}
\title{min diff}
\usage{
minDif(tVals, genes, ct)
}
\arguments{
\item{tVals}{tVals}

\item{genes}{vector of gene names}

\item{ct}{ct to compare to}
}
\value{
vector of differences
}
\description{
computes mean gene A in CT 1 - mean gene A in CT 2, where CT2 has the non CT1 max value. does this for genes
}
