% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmTFs}
\alias{mmTFs}
\title{mouse transcriptional regulators}
\format{A vector with 1511 gene names
\describe{ 
}}
\source{
org.Mm.eg.db and GO.db
}
\usage{
mmTFs
}
\description{
A vector of gene symbols representing transcription factors, defined as genes annotated with GO BP as 'regulation of transcription' and 'DNA binding'
}
\keyword{datasets}

