% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{mp_rainbowPlot}
\alias{mp_rainbowPlot}
\title{make a rainbow colored dot plot}
\usage{
mp_rainbowPlot(expDat, stAll, gene, dLevel = "description1")
}
\arguments{
\item{expDat}{expression data matrix}

\item{stAll}{sample table}

\item{gene}{gene name}

\item{dLevel}{column name to group samples by}
}
\value{
ggplot object
}
\description{
make a rainbow colored dot plot
}
\examples{
mp_rainbowPlot(expDat,sampTab, "Actb", "description1")

}
