% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{multiplot}
\alias{multiplot}
\title{Create multiplot - adapted from Bioconductor package "scater"}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\description{
Create multiplot - adapted from Bioconductor package "scater"
}
