% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{Norm_cleanExp}
\alias{Norm_cleanExp}
\title{summariz expression when there are multiple measures per gene in a matrix}
\usage{
Norm_cleanExp(expDat, geneTab, nameCol = "symbol")
}
\arguments{
\item{expDat}{exp matrix}

\item{geneTab}{gene annotation}

\item{nameCol}{how genes are uniquely IDed in geneTab}
}
\value{
expression matrix
}
\description{
returns the gene averaged matrix
}
