% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{Norm_cleanPropRaw}
\alias{Norm_cleanPropRaw}
\title{load and normalize each group separately, then return expProp (proportional expression)}
\usage{
Norm_cleanPropRaw(sampTab, platform, inc = 1000)
}
\arguments{
\item{sampTab}{sample table as returned by geo_fixNames}

\item{platform}{platform name. Options are ...}

\item{inc}{number of samples to process simultaneously}
}
\value{
raw expression expressed as a proportion of overall signal per sample
}
\description{
load and normalize each group separately, then return expProp (proportional expression)
}
