% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{Norm_loadAnnotation}
\alias{Norm_loadAnnotation}
\title{load the specified gene annotation table}
\usage{
Norm_loadAnnotation(pName)
}
\arguments{
\item{pName}{}
}
\value{
table of probe_id, entrez id and gene symbols
}
\description{
load the specified gene annotation table
}
