% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{Norm_preNormRaw}
\alias{Norm_preNormRaw}
\title{Run justRMA on samples, nor norm, just BG}
\usage{
Norm_preNormRaw(sampTab, inc = 1000)
}
\arguments{
\item{sampTab}{sample table}

\item{inc}{number of CEL files to load at one time}
}
\value{
cleaned eSet obj
}
\description{
Run justRMA on samples, nor norm, just BG
}
