% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{Norm_quantNorm}
\alias{Norm_quantNorm}
\title{quantile normalize expression matrix}
\usage{
Norm_quantNorm(expDat)
}
\arguments{
\item{expDat}{}
}
\value{
quantile normalized matrix
}
\description{
equalize the distributions of columns
}
