% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{ordered_class_perf_hm}
\alias{ordered_class_perf_hm}
\title{Output ordered classifier performance heatmap by description1 label of validation samples
Added 6-4-18}
\usage{
ordered_class_perf_hm(cnProc, classifierPerformance)
}
\arguments{
\item{cnProc}{cellNet processor object from training}

\item{classifierPerformance}{cellNet classifier performance object from training}
}
\value{
nothing
}
\description{
Output ordered classifier performance heatmap by description1 label of validation samples
Added 6-4-18
}
\examples{
ordered_class_perf_hm(cnProcFile="cnProc_HS_RS_Jun_20_2017.rda", classPerfFile="classifierPerformance_June-20-2017.rda")

}
