% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{pdf_classification_hm}
\alias{pdf_classification_hm}
\title{Added 6-4-18
Create pdf of classification heatmap, one column for every sample.}
\usage{
pdf_classification_hm(cnResQuery, study_name)
}
\arguments{
\item{cnResQuery}{CellNet query results object}

\item{study_name}{desired study name}
}
\description{
Added 6-4-18
Create pdf of classification heatmap, one column for every sample.
}
