% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{pdf_grn_status}
\alias{pdf_grn_status}
\title{Create pdf of GRN status of starting and target cell types, grouping samples by dlevel}
\usage{
pdf_grn_status(cnResQuery, cnProc, cell_type_1, cell_type_2, study_name,
  dlevel = "description2")
}
\arguments{
\item{cnResQuery}{CellNet query results object}

\item{cnProc}{CellNet Processor object from training}

\item{cell_type_1}{starting cell/tissue type}

\item{cell_type_2}{ending or target cell/tissue type}

\item{study_name}{desired study name}

\item{dlevel}{column of sample table used for sample grouping}
}
\description{
Create pdf of GRN status of starting and target cell types, grouping samples by dlevel
}
