% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{pdf_grn_status_by_CT}
\alias{pdf_grn_status_by_CT}
\title{Create pdf of GRN status plots, one plot with all dlevel groups for every C/T type}
\usage{
pdf_grn_status_by_CT(cnResQuery, cnProc, study_name,
  dlevel = "description2")
}
\arguments{
\item{cnResQuery}{CellNet query results object}

\item{cnProc}{CellNet Processor object from training}

\item{study_name}{desired study name}

\item{dlevel}{column of sample table used for sample grouping}
}
\description{
Create pdf of GRN status plots, one plot with all dlevel groups for every C/T type
}
