% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{pdf_nis_plots}
\alias{pdf_nis_plots}
\title{Create pdf of Network Influence Scores (NISs) for given C/T type, grouping samples by dlevel}
\usage{
pdf_nis_plots(cnResQuery, cnProc, target_cell_type, study_name,
  dlevel = "description2")
}
\arguments{
\item{cnResQuery}{CellNet query results object}

\item{cnProc}{CellNet Processor object from training}

\item{target_cell_type}{ending or target cell/tissue type}

\item{study_name}{desired study name}

\item{dlevel}{column of sample table used for sample grouping}
}
\description{
Create pdf of Network Influence Scores (NISs) for given C/T type, grouping samples by dlevel
}
