% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{plot_class_rocs}
\alias{plot_class_rocs}
\title{Plot results of cn_classAssess}
\usage{
plot_class_rocs(assessed)
}
\arguments{
\item{assessed}{result of runnung cn_classAssess}
}
\value{
ggplot pbject
}
\description{
Plot one ROC per CT
}
\examples{
testAssTues<-cn_splitMakeAssess(stTrain, expTrain, ctGRNs, prop=.5)
plot_class_rocs(testAssTues$ROCs)

}

