% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_graphics.R
\name{plot_nis}
\alias{plot_nis}
\title{boxplot of network influence scores}
\usage{
plot_nis(tfScores, targetCT, sampTab, group, limitTo = 20,
  dLevel = "description1")
}
\arguments{
\item{tfScores}{result of running cn_nis_all}

\item{targetCT}{what is the target cell type?}

\item{sampTab}{sample table to select subset to plot}

\item{group}{which subset to plot}

\item{limitTo}{number of TFs to show}

\item{dLevel}{level to group on}
}
\value{
ggplot object of box and whiskers
}
\description{
NIS<0 means that the expression of the TF is lower in query samples as compared to target cell type, and magnitude determined by number and dysregualtion of target genes
}
\examples{
plot_nis(tfScores, "hspc", sampTab, "day21")

}
