% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_assess.R
\name{pool_auPRs}
\alias{pool_auPRs}
\title{pool auprs}
\usage{
pool_auPRs(perfDF, rng = seq(-4, 10, by = 1))
}
\arguments{
\item{perfDF}{perfDF}

\item{rng}{seq -4 to 10}
}
\value{
df
}
\description{
pool auprs
}
