% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{processSRA}
\alias{processSRA}
\title{processes a SRA study}
\usage{
processSRA(sampTab, inc = 5, studyID = NA, bucket = "cellnet-rnaseq",
  startDir = "singlecell", finalLength = 40, subProp = 0.001,
  target = "mouse",
  gtfFile = "/media/ephemeral1/dat/ref/Mus_musculus.GRCm38.83.gtf",
  geneTabfname = "/media/ephemeral1/dat/ref/geneToTrans_Mus_musculus.GRCm38.80.exo_Jun_02_2015.R",
  salmonIndex = "/media/ephemeral1/dat/ref/MM_GRCh38.SalmonIndex.030816")
}
\arguments{
\item{sampTab}{sample table}

\item{inc}{increment size}

\item{studyID}{study id}

\item{bucket}{S3 bucket where to put the results}

\item{startDir}{S3 bucket path to store results}

\item{finalLength}{final length of reads}

\item{subProp}{proportion of reads to sample from for QC}

\item{target}{species/genome for expression estimates}

\item{gtfFile}{genomic feature of target genome for QC}

\item{geneTabfname}{transcript <-> gene data frame}

\item{salmonIndex}{path to salmon index for expression estimation}
}
\value{
sample table with QC measures appened. side effect is storage of expression estimates + sample table to S3
}
\description{
This function is a wrapper to qcAndSalmon. It also summarizes transcript into gene expression estimates
}
\examples{
processSRA(stX, inc=5)

}
