% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{qc_analysis}
\alias{qc_analysis}
\title{Run qc pipeline}
\usage{
qc_analysis(sampTab, sname = "sra_id", subProp = 0.001, target = "mouse",
  gtfFile = "/media/ephemeral1/dat/ref/Mus_musculus.GRCm38.83.gtf",
  cname = "trimNames")
}
\arguments{
\item{sampTab}{sample table}

\item{sname}{column of default prefix}

\item{subProp}{fraction of reads to sample}

\item{target}{species/genome}

\item{gtfFile}{gtf path/filename}

\item{cname}{column name containing input file names}
}
\value{
sample tablein which qc measures have been appended
}
\description{
subsamples reads, aligns to each genome, and counts the reads mapping to features in the target genome
}
