% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{qcAndSalmon}
\alias{qcAndSalmon}
\title{Runs the qc and salmon pipeline used by processSRA}
\usage{
qcAndSalmon(sampTab, studyID, bucket = "cellnet-rnaseq", finalLength = 40,
  subProp = 0.001, target = "mouse",
  gtfFile = "/media/ephemeral1/dat/ref/Mus_musculus.GRCm38.83.gtf",
  salmonIndex = "/media/ephemeral1/dat/ref/MM_GRCh38.SalmonIndex.030816")
}
\arguments{
\item{sampTab}{sample table}

\item{studyID}{study id}

\item{bucket}{S3 bucket where to put the results}

\item{finalLength}{final length of reads}

\item{subProp}{proportion of reads to sample from for QC}

\item{target}{species/genome for expression estimates}

\item{gtfFile}{genomic feature of target genome for QC}

\item{salmonIndex}{path to salmon index for expression estimation}
}
\value{
sample table with QC measures appened
}
\description{
Fetch raw sra files named in sampTab and convert to fastq (using SRA toolkit), trim reads to standard length, run QC by aligning sub-sampled fastq files aginst several genomes (human, mouse, zebrafish, fly, ecoli, yeast), and counting hits to genomic features of the target genome. Finish by estimating gene expression levels with Salmon.
}
