% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_utils.R
\name{s3_listDir}
\alias{s3_listDir}
\title{List directories in a bucket}
\usage{
s3_listDir(bucket, path)
}
\arguments{
\item{bucket}{bucket}

\item{path}{path}
}
\value{
vector of directories
}
\description{
List directories in a bucket
}
