% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_utils.R
\name{s3_put}
\alias{s3_put}
\title{put file from S3 using CLI}
\usage{
s3_put(dir, target, bucket = "pcahanrnaseq")
}
\arguments{
\item{dir}{path}

\item{target}{filename}

\item{bucket}{bucket}
}
\value{
nothing
}
\description{
put file from S3 using CLI, assumes AWS credentials are set
}
