% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{salmon_load_tranEst}
\alias{salmon_load_tranEst}
\title{load Salmon-based transcript estimates}
\usage{
salmon_load_tranEst(sampTab, cname = "sra_id", prefix = "salmonRes")
}
\arguments{
\item{sampTab}{sample table}

\item{cname}{used as convetion to name output files}

\item{prefix}{text prefix to output files}
}
\value{
list of exp matricies of both NumReads and TPM
}
\description{
load Salmon-based transcript estimates from quant.sf files
}
