% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{salmon_par}
\alias{salmon_par}
\title{quantify transcript levels using 'pseudo-alignments'}
\usage{
salmon_par(sampTab, sname = "sra_id",
  salmonIndex = "/media/ephemeral1/dat/ref/MM_GRCh38.SalmonIndex.030816",
  cname = "trimNames", libraryType = "U", numThreads = 5,
  destPrefix = "salmonRes", njobs = 4,
  salmonPath = "~/rnaseq/SalmonBeta-0.6.1_DebianSqueeze/bin")
}
\arguments{
\item{sampTab}{sample table}

\item{sname}{colname to id sample}

\item{salmonIndex}{salmon index directory}

\item{cname}{trimNames}

\item{libraryType}{library type (unstranded, etc)}

\item{numThreads}{number of threads to use}

\item{destPrefix}{where should the resulting files be placed}

\item{njobs}{use shell parallel too, and allot njobs}

\item{salmonPath}{path to salmon binary}
}
\value{
vector of output names
}
\description{
Uses salmon
}
