% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNet.R
\name{samp_for_class}
\alias{samp_for_class}
\title{split data into train vs test}
\usage{
samp_for_class(sampTab, prop = 0.5, dLevel = "description1",
  dLevelStudy = "exp_id")
}
\arguments{
\item{sampTab}{sample table must have exp_id column}

\item{prop}{proportion of samples to use as training}

\item{dLevel}{column name for}

\item{dLevelStudy}{column name to indicate experiment or study id}
}
\value{
list of stTrain stVal
}
\description{
Splits a sample table into 2 sample tables roughly by prop in which no samples with sampe exp_id are in both the test and train
}
