% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_grn.R
\name{sample_profiles_grn}
\alias{sample_profiles_grn}
\title{sample equivalent numbers of profiles per cell type}
\usage{
sample_profiles_grn(sampTab, minNum = NULL, dLevel = "description1")
}
\arguments{
\item{sampTab}{sample table}

\item{minNum}{min number of samples to get per CT}

\item{dLevel}{grouping}
}
\value{
subset of sample table
}
\description{
sample equivalent numbers of profiles per cell type
}
