% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{sra_breakTable}
\alias{sra_breakTable}
\title{break a sample table into _inc_ size pieces}
\usage{
sra_breakTable(sampTab, inc = 6)
}
\arguments{
\item{sampTab}{sample table}

\item{inc}{increment size}
}
\value{
list of sample tables
}
\description{
break a sample table into _inc_ size pieces. used in processSRA.
}
