% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{sra_delete}
\alias{sra_delete}
\title{Deletes files with specified suffix in specified directory}
\usage{
sra_delete(xdir = "/media/ephemeral0/dat/seq/tmpOut", suff = "gz")
}
\arguments{
\item{xdir}{path to files}

\item{suff}{what types of files to rm}
}
\value{
nothing
}
\description{
Deletes files with specified suffix in specified directory. Used in fetchAndConvert
}
