% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{sra_fetch}
\alias{sra_fetch}
\title{fetch sras to path defined when configuring sra tools.}
\usage{
sra_fetch(sra_ids)
}
\arguments{
\item{sra_ids}{sra accessions to fetch}
}
\value{
nothing
}
\description{
fetch sras to path defined when configuring sra tools. does it in parallel. for now limited to files <=600GB.
}
