% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{sra_parse}
\alias{sra_parse}
\title{unpacks the SRA files, splits them, and deletes SRAs}
\usage{
sra_parse(sra_ids, delete = TRUE, outdir = "/media/ephemeral0/dat",
  indir = "/media/ephemeral1/dat/seq/sra")
}
\arguments{
\item{sra_ids}{sra ids}

\item{delete}{whether or not to delete the originals after fastq-dump}

\item{outdir}{where to store resulting fastw files}

\item{indir}{where are the sras stored by sra-toolkit}
}
\value{
nothing
}
\description{
unpacks the SRA files, splits them, and deletes SRAs
}
