% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{subsamp_fastqs}
\alias{subsamp_fastqs}
\title{randomly samples reads from single end fastqs}
\usage{
subsamp_fastqs(fnames, codeDir = "~/code/", prop = 0.01)
}
\arguments{
\item{fnames}{files to subsample}

\item{path}{to python script subsetOne.py}

\item{proporation}{of reads to select}
}
\value{
nothing
}
\description{
randomly samples reads from single end fastqs uses python script in parallel
}
