% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellNet.R
\name{subSamp_for_class}
\alias{subSamp_for_class}
\title{select from sample table}
\usage{
subSamp_for_class(sampTab, prop = 0.5)
}
\arguments{
\item{sampTab}{sample table}

\item{prop}{fraction of samples to select}
}
\value{
stTrain
}
\description{
that's it. helper function
}

