% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{trans_dNorm}
\alias{trans_dNorm}
\title{express as a fraction of the column total}
\usage{
trans_dNorm(expDat, xFact = 1e+05)
}
\arguments{
\item{expDat}{expression matrix}

\item{xFact}{scale by this value}
}
\value{
transformed data matrix
}
\description{
express as a fraction of the column total
}
