% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{trans_eShift}
\alias{trans_eShift}
\title{shift the expression of each sample so that the min=minVal}
\usage{
trans_eShift(expDat, minVal = 0)
}
\arguments{
\item{expDat}{expression matrix}

\item{minVal}{}
}
\value{
transformed expression matrix
}
\description{
shift the expression of each sample so that the min=minVal
}
