% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_preprocess.R
\name{trans_fetch_index}
\alias{trans_fetch_index}
\title{get files needed to run Salmon and Hisat2 QC pipeline}
\usage{
trans_fetch_index(destination = "/media/ephemeral1/dat/ref",
  species = "mouse", bucket = "cellnet-rnaseq", dir = "ref")
}
\arguments{
\item{destination}{where to store them}

\item{species}{species. e.g. 'mouse'}

\item{bucket}{name of source s3 bucket}

\item{dir}{path on S3 bucket}
}
\value{
nothing
}
\description{
get files needed to run Salmon and Hisat2 QC pipeline from S#
}
