% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_utils.R
\name{utils_count_comments}
\alias{utils_count_comments}
\title{count comments in a file}
\usage{
utils_count_comments(fname)
}
\arguments{
\item{fname}{file}
}
\value{
numeric
}
\description{
count comments in a file, assumes #
}
