% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_utils.R
\name{utils_log}
\alias{utils_log}
\title{make a logging string}
\usage{
utils_log(str, prepend = "")
}
\arguments{
\item{str}{to log}

\item{str}{to prepend if wanted}
}
\value{
output str
}
\description{
make a logging string
}
