% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_utils.R
\name{utils_tarcompress_dir}
\alias{utils_tarcompress_dir}
\title{tars and compresses (gzip) dir}
\usage{
utils_tarcompress_dir(dir)
}
\arguments{
\item{dir}{directory}
}
\value{
file name
}
\description{
tars and compresses (gzip) dir
}
